fix(sandbox): recover a gateway wedged in any not-serving state
The in-sandbox serving watchdog only ever escalated on curl exit 7, so a
gateway whose process stayed alive while its listener became unusable was
recovered in exactly one shape: an unbroken streak of pure connection-refused
probes. Every other wedge signature read as "a listener exists" and silently
re-armed the watchdog:

  * curl 28 - the socket accepts but nothing answers within the timeout
  * curl 52/56 - accepted then dropped, which is the 1006 abnormal-closure
    transport error surfaced by 'openclaw health'
  * curl 0 with an HTTP error status - /health replies but no session serves

Those outcomes were delegated to the Docker HEALTHCHECK, but nothing restarts
an unhealthy OpenShell sandbox container, so the gateway stayed wedged. A
wedge that alternated between refused and any other failure also reset the
consecutive-refusal counter on every other probe and never reached the
threshold at all, leaving a live watchdog that logged nothing and recovered
nothing.

Classify each probe with the same definition the boot-time readiness gate
uses - /health must answer 200 or 401 - and count every not-serving outcome
toward one threshold. Probes that fail for local reasons (curl missing or
otherwise unable to run) stay inconclusive: they neither arm the watchdog nor
count toward the threshold, so a broken probe can never become a kill loop.
Each failed probe now logs the cause it observed.

Arming after the first serving probe, the PID-identity and cmdline checks
before signalling, and the respawn-marker handshake are unchanged. The
watchdog suite moves to its own file with the shared shell-harness helpers
extracted, keeping both inside the test-file size budget.

Fixes #7377
