+++ docs/reference/troubleshooting.mdx
@@ -639,7 +639,14 @@ Docker then marks the container `(unhealthy)` even though `pgrep` still finds th
 NemoClaw prevents and self-heals this:
 
 - The generated sandbox config pins `gateway.reload.mode` to `hot`, so configuration changes never make the gateway restart itself out from under the sandbox supervisor.
-- A serving watchdog inside the sandbox kills a gateway that stops listening after it has served, and the supervisor relaunches it (look for `[gateway-watchdog]` lines in `$$nemoclaw <name> logs`).
+- A serving watchdog inside the sandbox kills a gateway that stops serving after it has served at least once, and the supervisor relaunches it (look for `[gateway-watchdog]` lines in `$$nemoclaw <name> logs`).
+
+The watchdog treats the gateway as serving only while the local health endpoint answers `200` or `401`, the same check the sandbox uses at startup.
+A refused port, a probe that times out, a connection that is accepted and then dropped, and an HTTP error response all count as not serving, so a gateway wedged in any of those states is recovered rather than left running.
+The watchdog logs the cause it observed on each failed probe, and acts after four consecutive failures (roughly two minutes).
+
+A gateway wedged this way shows up as `openclaw health` reporting `gateway_transport_error` — often `1006 abnormal closure (no close frame)` — while the container stays `running` with no restarts, and agent replies keep working through the embedded fallback route instead of the gateway.
+If the watchdog reports `health probe inconclusive`, the probe itself could not run inside the sandbox; the gateway is left untouched, and the sandbox image needs repair with `$$nemoclaw <name> rebuild --yes`.
 
 Because of the `hot` pin, restart-class configuration changes made inside the sandbox log `config reload requires gateway restart; hot mode ignoring` and do not take effect until the gateway restarts.
 For example, `openclaw plugins install` logs that message until the gateway restarts.
