+++ test/nemoclaw-start-gateway-health.test.ts
@@ -2,12 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 // Gateway-health coverage for scripts/nemoclaw-start.sh (#4503, #4710):
-// the Docker HEALTHCHECK marker invariants and the gateway serving watchdog.
-// The OpenClaw gateway can drop its HTTP listener while the process stays
-// alive (failed in-process SIGUSR1 restart); the #2757 respawn loop only sees
-// process exit, so the watchdog must kill an alive-but-deaf gateway to hand
-// recovery back to the respawn loop. Marker tests are split from
-// test/nemoclaw-start.test.ts, which is at its size budget
+// the Docker HEALTHCHECK marker invariants and the supervised gateway
+// lifecycle. The serving watchdog that recovers an alive-but-not-serving
+// gateway is covered in test/gateway-serving-watchdog.test.ts. Marker tests
+// are split from test/nemoclaw-start.test.ts, which is at its size budget
 // (ci/test-file-size-budget.json).
 
 import { spawnSync } from "node:child_process";
@@ -16,90 +14,19 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { describe, expect, it } from "vitest";
 
-const START_SCRIPT = path.join(import.meta.dirname, "..", "scripts", "nemoclaw-start.sh");
-const GATEWAY_SUPERVISOR = path.join(
-  import.meta.dirname,
-  "..",
-  "scripts",
-  "lib",
-  "gateway-supervisor.sh",
-);
-
-// Read a file that may legitimately be absent without a check-then-read
-// race (CodeQL js/file-system-race): attempt the read and treat a missing
-// file as null.
-function readFileIfPresent(filePath: string): string | null {
-  try {
-    return fs.readFileSync(filePath, "utf-8");
-  } catch {
-    return null;
-  }
-}
-
-function extractShellFunction(src: string, name: string): string {
-  const header = `${name}() {`;
-  const start = src.indexOf(header);
-  expect(start, `Expected ${name} in scripts/nemoclaw-start.sh`).not.toBe(-1);
-  const bodyStart = start + header.length;
-  const body = src.slice(bodyStart);
-  const closing = body.match(/^}$/m);
-  expect(closing, `Expected closing brace for ${name} in scripts/nemoclaw-start.sh`).not.toBeNull();
-  return `${name}() {${body.slice(0, closing?.index ?? 0)}\n}`;
-}
+import {
+  extractGatewayLogAppendFunction,
+  extractShellFunction,
+  GATEWAY_SUPERVISOR,
+  pidIdentityFunctions,
+  readFileIfPresent,
+  START_SCRIPT,
+  safeTmpHelpers,
+  writeProcStatFunction,
+} from "./nemoclaw-start-gateway.test-helpers";
 
 function gatewayMarkerFunction(src: string, name: string, markerPath: string): string {
   return extractShellFunction(src, name).replaceAll("/tmp/nemoclaw-gateway-local", markerPath);
-}
-
-function extractGatewayLogAppendFunction(src: string, gatewayLog: string): string {
-  const functionSource = extractShellFunction(src, "append_openclaw_gateway_log_line");
-  const marker = '  local log_file="/tmp/gateway.log"';
-  expect(functionSource).toContain(marker);
-  return functionSource.replace(marker, `  local log_file=${JSON.stringify(gatewayLog)}`);
-}
-
-function safeTmpHelpers(src: string): string {
-  const start = src.indexOf("_nemoclaw_safe_replace_tmp_file() {");
-  const end = src.indexOf("_START_LOG=", Math.max(start, 0));
-  expect(start, "Expected safe temp helpers in scripts/nemoclaw-start.sh").not.toBe(-1);
-  expect(end, "Expected safe temp helpers in scripts/nemoclaw-start.sh").toBeGreaterThan(start);
-  return src.slice(start, end);
-}
-
-function pidIdentityFunctions(src: string): string {
-  const supervisor = fs.readFileSync(GATEWAY_SUPERVISOR, "utf-8");
-  return [
-    extractShellFunction(src, "openclaw_load_pid_identity"),
-    extractShellFunction(src, "openclaw_pid_start_identity"),
-    extractShellFunction(src, "capture_openclaw_pid_start_identity"),
-    extractShellFunction(src, "openclaw_supervised_pid_is_live"),
-    extractShellFunction(supervisor, "gateway_control_proc_root"),
-    extractShellFunction(supervisor, "gateway_control_proc_root_is_explicit"),
-    extractShellFunction(supervisor, "gateway_control_pid_state"),
-    extractShellFunction(supervisor, "gateway_control_pid_is_live"),
-  ].join("\n");
-}
-
-const writeProcStatFunction = [
-  "write_proc_stat() {",
-  '  local pid="$1" parent="$2" start="$3"',
-  '  printf \'%s (test-process) S %s\' "$pid" "$parent"',
-  "  for _ in {1..17}; do printf ' 0'; done",
-  "  printf ' %s\\n' \"$start\"",
-  "}",
-].join("\n");
-
-function watchdogFunctions(gatewayLog: string): string {
-  const src = fs.readFileSync(START_SCRIPT, "utf-8");
-  return [
-    safeTmpHelpers(src),
-    pidIdentityFunctions(src),
-    extractGatewayLogAppendFunction(src, gatewayLog),
-    extractShellFunction(src, "record_gateway_pid"),
-    extractShellFunction(src, "gateway_pid_is_openclaw_gateway"),
-    extractShellFunction(src, "gateway_watchdog_positive_int_ok"),
-    extractShellFunction(src, "start_gateway_serving_watchdog"),
-  ].join("\n");
 }
 
 function rootGatewayLifecycleFunctions(src: string, gatewayLog: string): string {
@@ -127,389 +54,6 @@ function gatewayLaunchBlock(src: string, kind: "non-root" | "root", gatewayLog: 
   expect(end, `Expected ${kind} gateway launch block in scripts/nemoclaw-start.sh`).not.toBe(-1);
   return src.slice(start, src.indexOf("\n", end)).replaceAll("/tmp/gateway.log", gatewayLog);
 }
-
-// Drive the watchdog end-to-end against a real background process standing in
-// for the gateway. `curlPlan` is the sequence of curl exit codes the stubbed
-// probe returns, one per watchdog cycle; the last entry repeats forever.
-// The proc fixture under _NEMOCLAW_PROC_ROOT controls what the PID-identity
-// check sees for the fake gateway.
-function runWatchdog(opts: {
-  curlPlan: number[];
-  cmdline?: string;
-  env?: Record<string, string>;
-  // How long to let the watchdog run when no kill is expected (seconds).
-  settleSeconds?: number;
-  expectKill: boolean;
-}): {
-  result: ReturnType<typeof spawnSync>;
-  fakeAlive: boolean;
-  wedgeLog: string;
-  tmpDir: string;
-} {
-  const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-watchdog-"));
-  const planFile = path.join(tmpDir, "curl-plan.txt");
-  const wedgeLogFile = path.join(tmpDir, "gateway.log");
-  fs.writeFileSync(wedgeLogFile, "", { mode: 0o644 });
-  const pidFile = path.join(tmpDir, "gateway.pid");
-  const procRoot = path.join(tmpDir, "proc");
-  fs.writeFileSync(planFile, `${opts.curlPlan.join("\n")}\n`);
-
-  const settle = opts.settleSeconds ?? 0.5;
-  const wrapper = [
-    "#!/usr/bin/env bash",
-    "set -o pipefail",
-    `GATEWAY_PID_FILE=${JSON.stringify(pidFile)}`,
-    "_DASHBOARD_PORT=18789",
-    `_NEMOCLAW_PROC_ROOT=${JSON.stringify(procRoot)}`,
-    `_NEMOCLAW_GATEWAY_LOG=${JSON.stringify(path.join(tmpDir, "ignored-inherited.log"))}`,
-    writeProcStatFunction,
-    // Throttle rather than no-op so the spinning loop stays cheap but the
-    // test still completes in well under a second per cycle.
-    "sleep() { command sleep 0.01; }",
-    // curl stub: pop the next exit code off the plan; keep the last one.
-    `_CURL_PLAN=${JSON.stringify(planFile)}`,
-    "curl() {",
-    "  local next rest",
-    '  next="$(head -n1 "$_CURL_PLAN" 2>/dev/null)"',
-    '  [ -n "$next" ] || next=0',
-    '  rest="$(tail -n +2 "$_CURL_PLAN" 2>/dev/null)"',
-    '  if [ -n "$rest" ]; then printf "%s\\n" "$rest" >"$_CURL_PLAN"; fi',
-    '  return "$next"',
-    "}",
-    // A real process stands in for the gateway so kill -0 / kill -TERM are
-    // exercised for real; its claimed cmdline comes from the proc fixture.
-    "command sleep 60 &",
-    "FAKE_GATEWAY_PID=$!",
-    "FAKE_GATEWAY_START=1001",
-    `mkdir -p ${JSON.stringify(procRoot)}/$FAKE_GATEWAY_PID`,
-    `printf '%s' ${JSON.stringify(opts.cmdline ?? "openclaw-gateway")} >${JSON.stringify(procRoot)}/$FAKE_GATEWAY_PID/cmdline`,
-    `write_proc_stat "$FAKE_GATEWAY_PID" "$$" "$FAKE_GATEWAY_START" >${JSON.stringify(procRoot)}/$FAKE_GATEWAY_PID/stat`,
-    watchdogFunctions(wedgeLogFile),
-    'capture_openclaw_pid_start_identity() { printf -v "$2" "%s" "watchdog-test"; }',
-    'record_gateway_pid "$FAKE_GATEWAY_PID" "$FAKE_GATEWAY_START"',
-    "start_gateway_serving_watchdog",
-    'printf "WATCHDOG_PID=%s\\n" "$GATEWAY_WATCHDOG_PID"',
-    ...(opts.expectKill
-      ? [
-          // Poll until the watchdog kills the fake gateway (or time out).
-          "for _ in $(command seq 1 300); do",
-          '  kill -0 "$FAKE_GATEWAY_PID" 2>/dev/null || break',
-          "  command sleep 0.02",
-          "done",
-        ]
-      : [`command sleep ${settle}`]),
-    'if kill -0 "$FAKE_GATEWAY_PID" 2>/dev/null; then printf "FAKE_ALIVE=1\\n"; else printf "FAKE_ALIVE=0\\n"; fi',
-    // Disown before killing: bash's asynchronous job-termination report
-    // includes the full job command text (the watchdog subshell body), which
-    // would pollute stderr assertions.
-    "disown -a 2>/dev/null || true",
-    'kill -KILL "$GATEWAY_WATCHDOG_PID" 2>/dev/null || true',
-    'kill -KILL "$FAKE_GATEWAY_PID" 2>/dev/null || true',
-    "command sleep 0.05",
-  ].join("\n");
-
-  const script = path.join(tmpDir, "run.sh");
-  fs.writeFileSync(script, wrapper, { mode: 0o755 });
-
-  const result = spawnSync("bash", [script], {
-    encoding: "utf-8",
-    timeout: 30000,
-    env: { ...process.env, ...(opts.env ?? {}) },
-  });
-
-  const stdout = typeof result.stdout === "string" ? result.stdout : "";
-  const fakeAlive = /^FAKE_ALIVE=1$/m.test(stdout);
-  const wedgeLog = readFileIfPresent(wedgeLogFile) ?? "";
-  return { result, fakeAlive, wedgeLog, tmpDir };
-}
-
-describe("gateway serving watchdog (#4710)", () => {
-  it("kills an alive-but-deaf gateway after sustained connection-refused and logs CRITICAL", () => {
-    const { result, fakeAlive, wedgeLog, tmpDir } = runWatchdog({
-      curlPlan: [0, 7, 7, 7, 7],
-      expectKill: true,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fakeAlive).toBe(false);
-      expect(result.stderr).toContain("dropped its HTTP listener on port 18789");
-      expect(wedgeLog).toContain("[gateway-watchdog] CRITICAL");
-      expect(wedgeLog).toContain("dropped its HTTP listener on port 18789");
-      expect(wedgeLog).toContain("(#4710)");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("never arms — and never kills — when the gateway has not served yet", () => {
-    // A gateway that is still booting (or failed to boot) refuses from the
-    // start; that case belongs to the respawn loop and the Docker
-    // HEALTHCHECK, not the watchdog.
-    const { result, fakeAlive, wedgeLog, tmpDir } = runWatchdog({
-      curlPlan: [7],
-      expectKill: false,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fakeAlive).toBe(true);
-      expect(result.stderr).not.toContain("dropped its HTTP listener on port 18789");
-      expect(wedgeLog).toBe("");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("resets the refused streak when a probe succeeds again", () => {
-    // Three refusals (below the threshold of four), recovery, three more —
-    // the streak must reset at each success and the gateway must survive.
-    const { result, fakeAlive, tmpDir } = runWatchdog({
-      curlPlan: [0, 7, 7, 7, 0, 7, 7, 7, 0],
-      expectKill: false,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fakeAlive).toBe(true);
-      expect(result.stderr).not.toContain("dropped its HTTP listener on port 18789");
-      expect(result.stderr).toContain("refused connection (1/4)");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("treats curl timeout and HTTP-error outcomes as listener-present", () => {
-    // curl 28 (timeout) and 22 (HTTP error) prove a listener exists; they
-    // arm the watchdog but never count toward the refused streak — a wedged
-    // listener that still accepts connections stays the HEALTHCHECK's call.
-    const { result, fakeAlive, tmpDir } = runWatchdog({
-      curlPlan: [28, 22, 28, 22, 28, 22, 28, 22],
-      expectKill: false,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fakeAlive).toBe(true);
-      expect(result.stderr).not.toContain("dropped its HTTP listener on port 18789");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("does not kill a PID whose cmdline no longer looks like the gateway", () => {
-    const { result, fakeAlive, tmpDir } = runWatchdog({
-      curlPlan: [0, 7, 7, 7, 7],
-      cmdline: "vim notes.txt",
-      expectKill: false,
-      settleSeconds: 1.2,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fakeAlive).toBe(true);
-      expect(result.stderr).toContain("no longer looks like the openclaw gateway");
-      expect(result.stderr).not.toContain("dropped its HTTP listener on port 18789");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("honors the refused-threshold env override", () => {
-    const { result, fakeAlive, tmpDir } = runWatchdog({
-      curlPlan: [0, 7, 7],
-      env: { NEMOCLAW_GATEWAY_WATCHDOG_REFUSED_THRESHOLD: "2" },
-      expectKill: true,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fakeAlive).toBe(false);
-      expect(result.stderr).toContain("2 consecutive refused probes");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("falls back to defaults when the env knobs are not positive integers", () => {
-    // A zero/garbage interval would busy-loop the probe; a zero threshold
-    // would kill on the first refusal. Both must be rejected with a warning
-    // while the watchdog keeps working on the defaults.
-    const { result, fakeAlive, tmpDir } = runWatchdog({
-      curlPlan: [0, 7, 7, 7, 7],
-      env: {
-        NEMOCLAW_GATEWAY_WATCHDOG_INTERVAL_SECONDS: "0",
-        NEMOCLAW_GATEWAY_WATCHDOG_REFUSED_THRESHOLD: "banana",
-      },
-      expectKill: true,
-    });
-    try {
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(result.stderr).toContain(
-        "invalid NEMOCLAW_GATEWAY_WATCHDOG_INTERVAL_SECONDS='0'; defaulting to 30",
-      );
-      expect(result.stderr).toContain(
-        "invalid NEMOCLAW_GATEWAY_WATCHDOG_REFUSED_THRESHOLD='banana'; defaulting to 4",
-      );
-      // Default threshold of 4 still applies.
-      expect(fakeAlive).toBe(false);
-      expect(result.stderr).toContain("4 consecutive refused probes");
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("does not inherit the armed state when the pidfile switches to a new gateway PID", () => {
-    // A fast respawn can replace the pidfile between probes without the
-    // watchdog ever observing the old PID as dead. The new gateway must earn
-    // its own armed state — otherwise its boot-time refusals would count
-    // against the predecessor's serve history and it could be killed while
-    // still starting up.
-    const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-watchdog-swap-"));
-    try {
-      const planFile = path.join(tmpDir, "curl-plan.txt");
-      const probeLog = path.join(tmpDir, "probes.log");
-      const wedgeLogFile = path.join(tmpDir, "gateway.log");
-      fs.writeFileSync(wedgeLogFile, "", { mode: 0o644 });
-      const pidFile = path.join(tmpDir, "gateway.pid");
-      const procRoot = path.join(tmpDir, "proc");
-      // First probe arms on gateway A; everything after refuses.
-      fs.writeFileSync(planFile, "0\n7\n");
-
-      const wrapper = [
-        "#!/usr/bin/env bash",
-        "set -o pipefail",
-        `GATEWAY_PID_FILE=${JSON.stringify(pidFile)}`,
-        "_DASHBOARD_PORT=18789",
-        `_NEMOCLAW_PROC_ROOT=${JSON.stringify(procRoot)}`,
-        `_NEMOCLAW_GATEWAY_LOG=${JSON.stringify(path.join(tmpDir, "gateway.log"))}`,
-        writeProcStatFunction,
-        // A low threshold makes an inherited armed state lethal within a few
-        // cycles, so survival proves the per-PID reset.
-        "export NEMOCLAW_GATEWAY_WATCHDOG_REFUSED_THRESHOLD=2",
-        "sleep() { command sleep 0.01; }",
-        `_CURL_PLAN=${JSON.stringify(planFile)}`,
-        "curl() {",
-        "  local next rest",
-        '  next="$(head -n1 "$_CURL_PLAN" 2>/dev/null)"',
-        '  [ -n "$next" ] || next=0',
-        '  rest="$(tail -n +2 "$_CURL_PLAN" 2>/dev/null)"',
-        '  if [ -n "$rest" ]; then printf "%s\\n" "$rest" >"$_CURL_PLAN"; fi',
-        `  printf 'probe\\n' >> ${JSON.stringify(probeLog)}`,
-        '  case "$next" in',
-        '    0) record_gateway_pid "$GATEWAY_B" "$GATEWAY_B_START" ;;',
-        "  esac",
-        '  return "$next"',
-        "}",
-        "command sleep 60 &",
-        "GATEWAY_A=$!",
-        "command sleep 60 &",
-        "GATEWAY_B=$!",
-        "GATEWAY_A_START=2001",
-        "GATEWAY_B_START=2002",
-        `mkdir -p ${JSON.stringify(procRoot)}/$GATEWAY_A ${JSON.stringify(procRoot)}/$GATEWAY_B`,
-        `printf 'openclaw-gateway' >${JSON.stringify(procRoot)}/$GATEWAY_A/cmdline`,
-        `printf 'openclaw-gateway' >${JSON.stringify(procRoot)}/$GATEWAY_B/cmdline`,
-        `write_proc_stat "$GATEWAY_A" "$$" "$GATEWAY_A_START" >${JSON.stringify(procRoot)}/$GATEWAY_A/stat`,
-        `write_proc_stat "$GATEWAY_B" "$$" "$GATEWAY_B_START" >${JSON.stringify(procRoot)}/$GATEWAY_B/stat`,
-        watchdogFunctions(wedgeLogFile),
-        'capture_openclaw_pid_start_identity() { printf -v "$2" "%s" "watchdog-test"; }',
-        'record_gateway_pid "$GATEWAY_A" "$GATEWAY_A_START"',
-        "start_gateway_serving_watchdog",
-        // The curl stub swaps to gateway B during A's successful probe,
-        // before the watchdog can start counting refused probes again.
-        'printf "B_PID=%s\\n" "$GATEWAY_B"',
-        "command sleep 0.6",
-        'if kill -0 "$GATEWAY_B" 2>/dev/null; then printf "B_ALIVE=1\\n"; else printf "B_ALIVE=0\\n"; fi',
-        "disown -a 2>/dev/null || true",
-        'kill -KILL "$GATEWAY_WATCHDOG_PID" "$GATEWAY_A" "$GATEWAY_B" 2>/dev/null || true',
-        "command sleep 0.05",
-      ].join("\n");
-
-      const script = path.join(tmpDir, "run.sh");
-      fs.writeFileSync(script, wrapper, { mode: 0o755 });
-      const result = spawnSync("bash", [script], { encoding: "utf-8", timeout: 30000 });
-
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      const stdout = typeof result.stdout === "string" ? result.stdout : "";
-      // Without the per-PID reset, B inherits armed=1 and dies after two
-      // refused probes (threshold 2, 10ms cycles) well inside the 600ms
-      // observation window.
-      expect(stdout).toContain("B_ALIVE=1");
-      const bPid = stdout.match(/^B_PID=(\d+)$/m)?.[1];
-      expect(bPid).toBeDefined();
-      expect(result.stderr).not.toContain(
-        `gateway pid ${bPid} is alive but dropped its HTTP listener on port 18789`,
-      );
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-
-  it("tracks the replacement before a termination signal interrupts restart health wait", () => {
-    const src = fs.readFileSync(START_SCRIPT, "utf-8");
-    const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-restart-signal-race-"));
-    const eventLog = path.join(tmpDir, "events.log");
-    const scriptPath = path.join(tmpDir, "run.sh");
-
-    fs.writeFileSync(
-      scriptPath,
-      [
-        "#!/usr/bin/env bash",
-        "set -euo pipefail",
-        `EVENT_LOG=${JSON.stringify(eventLog)}`,
-        "GATEWAY_PID=101",
-        "GATEWAY_PID_START_IDENTITY=old-start",
-        "SANDBOX_WAIT_PID=101",
-        "SANDBOX_CHILD_PIDS=(101)",
-        "GATEWAY_CONTROL_ACTION=restart",
-        "GATEWAY_CONTROL_SIGNAL_PENDING=0",
-        "gateway_control_take_request() { :; }",
-        'openclaw_supervised_pid_is_live() { case "$1:$2" in "101:old-start"|"202:new-start") return 0 ;; *) return 1 ;; esac; }',
-        "gateway_control_stop_tracked_pid() {",
-        '  printf "stop:%s:%s\\n" "$1" "$2" >>"$EVENT_LOG"',
-        "  return 0",
-        "}",
-        "prepare_openclaw_gateway_restart() { :; }",
-        "run_openclaw_config_guard() { :; }",
-        "restore_openclaw_restart_config() { :; }",
-        "cleanup_openclaw_gateway_locks() { :; }",
-        "launch_openclaw_gateway() {",
-        "  GATEWAY_PID=202",
-        "  GATEWAY_PID_START_IDENTITY=new-start",
-        "  SANDBOX_WAIT_PID=202",
-        "}",
-        "wait_for_openclaw_gateway_internal() {",
-        '  kill -TERM "$$"',
-        "  return 1",
-        "}",
-        "start_plugin_registry_refresh() { :; }",
-        "gateway_control_complete() { :; }",
-        "gateway_control_fail() { :; }",
-        "cleanup_on_signal() {",
-        '  printf "cleanup:wait=%s:children=%s\\n" "$SANDBOX_WAIT_PID" "${SANDBOX_CHILD_PIDS[*]}" >>"$EVENT_LOG"',
-        '  [ "$SANDBOX_WAIT_PID" -eq 202 ]',
-        '  [ "${SANDBOX_CHILD_PIDS[*]}" = "202" ]',
-        "  exit 0",
-        "}",
-        "trap cleanup_on_signal SIGTERM SIGINT",
-        extractShellFunction(src, "openclaw_supervised_aux_pid_is_live"),
-        extractShellFunction(src, "stop_openclaw_supervised_gateway"),
-        extractShellFunction(src, "refresh_openclaw_supervised_child_pids"),
-        extractShellFunction(src, "mark_openclaw_gateway_stopped"),
-        extractShellFunction(src, "stop_openclaw_gateway_fail_closed"),
-        extractShellFunction(src, "retire_openclaw_supervised_gateway"),
-        extractShellFunction(src, "handle_openclaw_gateway_control_request"),
-        "handle_openclaw_gateway_control_request",
-      ].join("\n"),
-      { mode: 0o700 },
-    );
-
-    try {
-      const result = spawnSync("bash", [scriptPath], { encoding: "utf-8", timeout: 5000 });
-      expect(result.status, `script failed: ${result.stderr}`).toBe(0);
-      expect(fs.readFileSync(eventLog, "utf-8")).toBe(
-        "stop:101:old-start\ncleanup:wait=202:children=202\n",
-      );
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
-});
 
 describe("OpenClaw supervised child PID identity", () => {
   it("does not re-admit a recycled plugin-refresh PID owned by another process", () => {
